sfn-project/requirements.txt
asgiref
certifi
charset-normalizer
Django
django-extensions
django-environ
django-rest-framework
idna
pip
psycopg2-binary
pytz
requests
setuptools
sqlparse
tzdata
urllib3
wheel
gunicorn
django-heroku
